% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bdloss}
\alias{bdloss}
\title{Best estimates of the proportions of species threatened with extinction in
the Red List Index, by species group, 2021}
\format{
'bdloss'
A data frame with 15 rows and 3 columns:
\describe{
\item{Type}{Species type}
\item{Status}{Threat status}
\item{Percentage}{Percentage}
}
}
\source{
UN Sustainable Development Goal 15

\url{https://unstats.un.org/sdgs/report/2021/goal-15/}
}
\usage{
bdloss
}
\description{
This is an example dataset. Its helpful to have some data to play with when
testing the \link{ukfsr} package.
}
\keyword{datasets}
