% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bucket_manifest}
\alias{bucket_manifest}
\title{Get contents of the UKFSR S3 bucket}
\usage{
bucket_manifest(bucket = ukfsr::s3_bucket(), file_ext = "png")
}
\arguments{
\item{bucket}{An S3 bucket}

\item{file_ext}{A file extension}
}
\value{
A dataframe of the contents of the bucket, with indicator ids and
descriptions parsed from the filenames
}
\description{
Get contents of the UKFSR S3 bucket
}
\examples{
\dontrun{
x <- bucket_manifest(file_ext = "csv")
}
}
