% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{check_indicator}
\alias{check_indicator}
\title{Test whether a string is a valid UKFSR indicator id}
\usage{
check_indicator(string)
}
\arguments{
\item{string}{String containing a UKFSR indicator id}
}
\value{
A logical
}
\description{
Test whether a string is a valid UKFSR indicator id
}
\examples{
check_indicator("2.3.5a")
check_indicator("10.3.1")
}
