% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prodsupply}
\alias{prodsupply}
\title{UK production:supply ratio for food}
\format{
'prodsupply'
A data frame with 134 rows and 3 columns:
\describe{
\item{year}{Year}
\item{type}{Food category}
\item{value}{Production:supply ratio (\%)}
}
}
\source{
Agriculture in the UK:

\url{https://www.gov.uk/government/collections/agriculture-in-the-united-kingdom}
}
\usage{
prodsupply
}
\description{
This is an example dataset. Its helpful to have some data to play with when
testing the \link{ukfsr} package.
}
\keyword{datasets}
