% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{s3_bucket}
\alias{s3_bucket}
\title{Location of the UKFSR2024 Amazon S3 bucket}
\usage{
s3_bucket()
}
\value{
A string containing the bucket location
}
\description{
The bucket is used to store final publication data and graphics, as well as
underlying raw data.
}
\examples{
s3_bucket()
}
