% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_csv}
\alias{save_csv}
\title{Save UKFSR data for publication to the S3 bucket}
\usage{
save_csv(data, indicator_id, indicator_desc = "")
}
\arguments{
\item{data}{A data frame}

\item{indicator_id}{A valid UKFSR indicator id}

\item{indicator_desc}{A title for the graphic}
}
\value{
Saves a csv file to the S3 bucket
}
\description{
data is saved as a csv into a folder structure on the bucket as specified in
the UKFSR guidance. This is currently 'theme_x/tT_S_I/output/graphics'.
Filenames are derived from the indicator id and description. The bucket
location is encoded in \code{\link[=s3_bucket]{s3_bucket()}}.
}
\examples{
\dontrun{
save_csv(mtcars, "1.1.1", "mtcars data")
}

}
