% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukfsr-package.R
\docType{package}
\name{ukfsr-package}
\alias{ukfsr}
\alias{ukfsr-package}
\title{ukfsr: Tools to assist with producing the UK Food Security Report}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: David Lee \email{david.lee@defra.gov.uk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
