% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{check_indicator}
\alias{check_indicator}
\title{Test whether a string is a valid UKFSR indicator id}
\usage{
check_indicator(string)
}
\arguments{
\item{string}{String containing a UKFSR indicator id}
}
\value{
A logical
}
\description{
A UKFSR indicator can be uniquely identified from 3 main hierarchical components separated by periods:
\itemize{
\item theme (1-5)
\item section (1-9)
\item indicator (1-99)
}

with an optional variant suffix in the form (a -z). Thus a valid indicator id
is a string in the form "T.S.Iv". Some examples are "1.3.7a". "4.2.11".
check_indicator tests an indicator id string for compliance and returns TRUE
or FALSE.
}
\examples{
# A valid id
check_indicator("2.3.5a")

# An invalid id
check_indicator("10.3.1")
}
