% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{check_indicator}
\alias{check_indicator}
\title{Test whether a string is a valid UKFSR indicator id}
\usage{
check_indicator(string)
}
\arguments{
\item{string}{String containing a UKFSR indicator id}
}
\value{
A logical
}
\description{
UKFSR indicator ids contain 4 components:
}
\details{
\itemize{
\item theme (1-5)
\item section (1-9)
\item indicator (1-99)
\item variant - optional (a-z)
}

in the form \code{T.S.Iv}.
}
\examples{
check_indicator("2.3.5a")
check_indicator("10.3.1")
}
