% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_graphic}
\alias{save_graphic}
\title{Save a UKFSR graphic to the S3 bucket}
\usage{
save_graphic(graphic, indicator_id, indicator_desc = "")
}
\arguments{
\item{graphic}{A \code{ggplot} chart object}

\item{indicator_id}{A valid UKFSR indicator id}

\item{indicator_desc}{A title for the graphic}
}
\value{
Saves a png and svg file to the S3 bucket
}
\description{
Save a UKFSR graphic to the S3 bucket
}
\examples{
\dontrun{
chart <- ggplot2::ggplot(mtcars) + 
ggplot2::geom_point(ggplot2::aes(x = mpg, y = wt))

save_graphic(chart, "1.1.1", "Test graphic")
}
}
