% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_ukfsr}
\alias{theme_ukfsr}
\title{UKFSR chart theme}
\usage{
theme_ukfsr(
  base_size = 10,
  base_family = "",
  base_line_size = base_size/20,
  base_rect_size = base_size/20
)
}
\arguments{
\item{base_size}{Base font size in pts}

\item{base_family}{Base font family}

\item{base_line_size}{Base size for line elements}

\item{base_rect_size}{Base size for rect elements}
}
\description{
UKFSR chart theme
}
