% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_ukfsr}
\alias{theme_ukfsr}
\title{UKFSR chart theme}
\usage{
theme_ukfsr(
  base_size = 14,
  base_family = "",
  base_line_size = base_size/20,
  base_rect_size = base_size/20,
  horizontal = FALSE,
  chart_line_size = 2
)
}
\arguments{
\item{base_size}{Base font size in pts}

\item{base_family}{Base font family}

\item{base_line_size}{Base size for line elements}

\item{base_rect_size}{Base size for rect elements}

\item{horizontal}{Set to \code{TRUE} to adjust the theme for horizontal plots}

\item{chart_line_size}{sets the default linewidth for geom_line}
}
\value{
A chart theme
}
\description{
UKFSR chart theme
}
\section{Default theme settings}{
Used with the default parameters, \code{\link[=theme_ukfsr]{theme_ukfsr()}} formats charts in the
following way: the base_size defines the default text size at \strong{14pt}.
Other chart text elements are assigned font sizes relative to that using a
multiplier, as below.
\itemize{
\item \emph{axis.text:} 1.8
\item \emph{axis.title:} 1.8
\item \emph{legend.text:} 2
\item facet titles \emph{strip.text:} 2.2
\item \emph{plot.title} 1.2
\item \emph{plot.caption} 0.8
}

The default linewidth for lines and rectangles is \strong{0.5}.
}

\examples{
ggplot2::ggplot(mtcars) + 
ggplot2::geom_point(ggplot2::aes(x = mpg, y = wt)) + 
theme_ukfsr()
}
