% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_ukfsr}
\alias{theme_ukfsr}
\title{UKFSR chart theme}
\usage{
theme_ukfsr(
  base_size = 14,
  base_family = "",
  base_line_size = base_size/20,
  base_rect_size = base_size/20,
  horizontal = FALSE,
  chart_line_size = 2
)
}
\arguments{
\item{base_size}{Base font size in pts}

\item{base_family}{Base font family}

\item{base_line_size}{Base size for line elements}

\item{base_rect_size}{Base size for rect elements}

\item{horizontal}{Set to \code{TRUE} to adjust the theme for horizontal plots}

\item{chart_line_size}{sets the default line width for geom_line. See
\emph{Default theme settings}}
}
\value{
A chart theme
}
\description{
UKFSR chart theme
}
\section{Default theme settings}{
Used with the default parameters, \code{\link[=theme_ukfsr]{theme_ukfsr()}} formats charts in the
following way: the base_size defines the default text size at \strong{14pt}.
Other chart text elements are assigned font sizes relative to that using a
multiplier, as below.
\itemize{
\item \emph{axis.text:} 1.8
\item \emph{axis.title:} 1.8
\item \emph{legend.text:} 2
\item \emph{strip.text:} 2.2 (facet titles)
\item \emph{plot.title} 1.2
\item \emph{plot.caption} 0.8
}

The default linewidth for lines and rectangles is \strong{0.7}.

\strong{Chart line width}
\code{\link[=theme_ukfsr]{theme_ukfsr()}} uses \code{\link[ggplot2:update_defaults]{ggplot2::update_geom_defaults()}} to change the default
line width used by \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}. I'm not sure if this is strictly
speaking good practice to have a theme affect a geom. It will impact on any
uses of geom_line after \code{\link[=theme_ukfsr]{theme_ukfsr()}} has been applied. You can override in
individual charts with the lwd parameter.
}

\examples{
p <- ggplot2::ggplot(ukfsr::prodsupply) + 
ggplot2::geom_line(ggplot2::aes(x = year, y = value, colour = type))

p

p + theme_ukfsr()

b <- ggplot2::ggplot(ukfsr::bdloss) + 
ggplot2::geom_col(ggplot2::aes(x = Type, y = Percentage, fill = Status),
                  position = ggplot2::position_stack()) + 
ggplot2::coord_flip()

b

b + theme_ukfsr(base_size = 10, horizontal = TRUE)

}
