% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukfsr-package.R
\docType{package}
\name{ukfsr-package}
\alias{ukfsr}
\alias{ukfsr-package}
\title{ukfsr: Tools to assist with producing the UK Food Security Report}
\description{
Contains miscellaneous functions to help with wrangling data and graphics for the UKFSR.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://foodchainstats.github.io/ukfsr/}
}

}
\author{
\strong{Maintainer}: David Lee \email{david.lee@defra.gov.uk}

}
\keyword{internal}
